chore(Queries): remove query result pagination [YTFRONT-5766]
