--- packages/ui/src/ui/pages/query-tracker/QueryResultsView/ResultPaginator/index.scss
@@ -1,7 +1,0 @@
-.query-result-paginator {
-    display: inline-flex;
-
-    &__control:not(:last-child) {
-        margin-right: 4px;
-    }
-}
